backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
yara-python
onnxruntime
lightgbm
xgboost
numpy
python-magic
python-multipart
python-jose
uvicorn
